+++ tests/common/mod.rs
@@ -13,6 +13,7 @@ impl Register {
     pub const WTIME    : u8 = 0x03;
     pub const AILTL    : u8 = 0x04;
     pub const AIHTL    : u8 = 0x06;
+    pub const APERS    : u8 = 0x0C;
     pub const CONFIG   : u8 = 0x0D;
     pub const CONTROL  : u8 = 0x0F;
     pub const ID       : u8 = 0x12;